Add Vagrant based testing fixture (#24249)
