+++ plugins/jvm-example/build.gradle
@@ -33,7 +33,7 @@ dependencies {
   exampleFixture project(':test:fixtures:example-fixture')
 }
 
-task exampleFixture(type: org.elasticsearch.gradle.test.Fixture) {
+task exampleFixture(type: org.elasticsearch.gradle.test.AntFixture) {
   dependsOn project.configurations.exampleFixture
   executable = new File(project.javaHome, 'bin/java')
   args '-cp', "${ -> project.configurations.exampleFixture.asPath }",
